*************************************************** okay so let me explain what we got going on here. a project of mine that splits audio supplied as audio files into frequency files. pretty cool. now im implementing a neural-network that can split audio into stems.
as you can see, this is a new repo. my last repo got fucked mega and I
had to make a new one. I used demucs as an onnx standard in my onnx
runtime. i got it to work, very terribly, but my cuda gpu process was
fuckin nice!!! now im trying to do all of that but in the native runtime
of demucs..... pytorch!!!!

so thats whats been going on. back to the usual.

*COMPLETED*
- brought in libtorch library and friends

*TO-DO*
- get libtorch to connect with cuda.